backend/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn==0.30.0

# Database
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.0

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

# HTTP Client
httpx==0.27.0
requests==2.31.0

# Testing
pytest==8.0.0

# Supabase
supabase==2.3.0
python-dotenv==1.0.1

# Background Tasks & Caching
celery==5.3.6
redis==5.0.2

# Utilities
python-slugify==8.0.4
pillow==10.2.0
